russ: -- fix http inspect use of decompress_swf and decompress_pdf -- ensure that autotools and cmake install the same files in the same dirs -- add doc/online_manual.sh to generate an all-in-1 html manual with    embedded images

+++ src/service_inspectors/http_inspect/hi_ui_config.h
@@ -157,7 +157,7 @@ struct HTTPINSPECT_CONF
     char log_hostname;
     bool unlimited_decompress;
     char extract_gzip;
-    unsigned long file_decomp_modes;
+    uint32_t file_decomp_modes;
 
 #define HI_UI_CONFIG_XFF_FIELD_NAME  "X-Forwarded-For"
 #define HI_UI_CONFIG_TCI_FIELD_NAME  "True-Client-IP"
